requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.12.2
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.6.0
h11==0.16.0
html2text==2020.1.16
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.4.4
openai==1.82.1
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
